// CareerDevelopmentHubCode/apps/career-development-hub/src/generated/models/interaction-model.ts
import type { NetworkingContact } from './networking-contact-model';
import type { JobApplication } from './job-application-model';

export const InteractionInteractionTypeKeyToLabel = {
  'NetworkingChat': 'Networking Chat',
  'LinkedIn': 'LinkedIn',
  'Email': 'Email',
  'Call': 'Call',
  'Meeting': 'Meeting',
  'Event': 'Event',
  'Interview': 'Interview',
  'Other': 'Other'
} as const;
export type InteractionInteractionTypeKey = keyof typeof InteractionInteractionTypeKeyToLabel;

export interface Interaction {
  /**
   * @displayName Interaction
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Interaction Name
   * @description Human-readable record identifier
   */
  interactionName: string;
  /**
   * @displayName Contact
   * @validationRule Required for create/update operations
   */
  contact: Pick<NetworkingContact, 'id' | 'contactName'>;
  /**
   * @displayName Interaction Date
   * @validationRule Required for create/update operations
   */
  interactionDate: string;
  /**
   * @displayName Interaction Type
   * @validationRule Required for create/update operations
   */
  interactionTypeKey: InteractionInteractionTypeKey;
  /**
   * @displayName Notes
   */
  notes?: string;
  /**
   * @displayName Related Application
   */
  relatedApplication?: Pick<JobApplication, 'id' | 'role'>;
}

export const _Interaction = 'Interaction' as const;